+++ src/renderers/boolean.ts
@@ -1,6 +1,6 @@
 import figures from '@inquirer/figures'
 import { bgGray, bold, underline, white } from 'yoctocolors'
-import type { BooleanField } from '../types.js'
+import type { BooleanField } from '../util/types.js'
 
 export function renderBoolean(field: BooleanField, isFieldSelected: boolean): string {
     const { value } = field
