+++ src/util/stdout.ts
